refactor: strip CFR decompiler artifacts (banners, redundant casts, awkward control flow)
Made-with: Cursor

+++ src/main/java/group/austale/deitylandprotection/DeityLandProtectionBorderTickSystem.java
@@ -1,20 +1,3 @@
-/*
- * Decompiled with CFR 0.152.
- *
- * Could not load the following classes:
- *  com.hypixel.hytale.component.ArchetypeChunk
- *  com.hypixel.hytale.component.CommandBuffer
- *  com.hypixel.hytale.component.Ref
- *  com.hypixel.hytale.component.Store
- *  com.hypixel.hytale.component.query.Query
- *  com.hypixel.hytale.component.system.tick.EntityTickingSystem
- *  com.hypixel.hytale.math.vector.Transform
- *  com.hypixel.hytale.math.vector.Vector3d
- *  com.hypixel.hytale.protocol.Color
- *  com.hypixel.hytale.server.core.universe.PlayerRef
- *  com.hypixel.hytale.server.core.universe.world.ParticleUtil
- *  com.hypixel.hytale.server.core.universe.world.storage.EntityStore
- */
 package group.austale.deitylandprotection;
 
 import group.austale.deitylandprotection.Claim;
@@ -198,6 +181,6 @@ extends EntityTickingSystem<EntityStore> {
         }
         Vector3d pos = new Vector3d((double)x + 0.5, y, (double)z + 0.5);
         String id = particleSystemId == null || particleSystemId.isEmpty() ? SLUMBERING_PARTICLE_SYSTEM_ID : particleSystemId;
-        ParticleUtil.spawnParticleEffect((String)id, (Vector3d)pos, (float)0.0f, (float)0.0f, (float)0.0f, PARTICLE_SCALE, PARTICLE_COLOR, viewers, store);
+        ParticleUtil.spawnParticleEffect((String)id, (Vector3d)pos, 0.0f, 0.0f, 0.0f, PARTICLE_SCALE, PARTICLE_COLOR, viewers, store);
     }
 }
